imgui integration, deltatime originally was inputted as an int, which turned it into a zero, which made it useless, fixed now.

+++ src/Renderer.hpp
@@ -36,7 +36,7 @@ public:
     bool shouldClose();
     void swapBuffers();
 
-    void loop(std::function<void(int)>);
+    void loop(std::function<void(float)>);
 
     Renderer();
     ~Renderer();
